+++ portal/src/app/components/protocol-query/constants/algorithms-sections.ts
@@ -1,0 +1,1 @@
+export const algorithmSections = ['All', 'Classic', 'Hybrid', 'PQ'];
