+++ portal/src/app/components/protocol-query/constants/index.ts
@@ -1,0 +1,1 @@
+export * from './algorithms-sections';
